+++ src/main/java/com/imsweb/layout/LayoutFactory.java
@@ -46,58 +46,63 @@ import com.imsweb.layout.record.fixed.naaccr.NaaccrLayout;
 public final class LayoutFactory {
 
     //XML Layouts - constants for the internal layout IDs
-    public static final String LAYOUT_ID_NAACCR_XML_21 = "naaccr-xml-21"; // abstract for full abstract
+    public static final String LAYOUT_ID_NAACCR_XML_22 = "naaccr-xml-22"; // shortcut for full abstract
+    public static final String LAYOUT_ID_NAACCR_XML_22_ABSTRACT = "naaccr-xml-22-abstract";
+    public static final String LAYOUT_ID_NAACCR_XML_22_MODIFIED = "naaccr-xml-22-modified";
+    public static final String LAYOUT_ID_NAACCR_XML_22_CONFIDENTIAL = "naaccr-xml-22-confidential";
+    public static final String LAYOUT_ID_NAACCR_XML_22_INCIDENCE = "naaccr-xml-22-incidence";
+    public static final String LAYOUT_ID_NAACCR_XML_21 = "naaccr-xml-21"; // shortcut for full abstract
     public static final String LAYOUT_ID_NAACCR_XML_21_ABSTRACT = "naaccr-xml-21-abstract";
     public static final String LAYOUT_ID_NAACCR_XML_21_MODIFIED = "naaccr-xml-21-modified";
     public static final String LAYOUT_ID_NAACCR_XML_21_CONFIDENTIAL = "naaccr-xml-21-confidential";
     public static final String LAYOUT_ID_NAACCR_XML_21_INCIDENCE = "naaccr-xml-21-incidence";
-    public static final String LAYOUT_ID_NAACCR_XML_18 = "naaccr-xml-18"; // abstract for full abstract
+    public static final String LAYOUT_ID_NAACCR_XML_18 = "naaccr-xml-18"; // shortcut for full abstract
     public static final String LAYOUT_ID_NAACCR_XML_18_ABSTRACT = "naaccr-xml-18-abstract";
     public static final String LAYOUT_ID_NAACCR_XML_18_MODIFIED = "naaccr-xml-18-modified";
     public static final String LAYOUT_ID_NAACCR_XML_18_CONFIDENTIAL = "naaccr-xml-18-confidential";
     public static final String LAYOUT_ID_NAACCR_XML_18_INCIDENCE = "naaccr-xml-18-incidence";
-    public static final String LAYOUT_ID_NAACCR_XML_16 = "naaccr-xml-16"; // abstract for full abstract
+    public static final String LAYOUT_ID_NAACCR_XML_16 = "naaccr-xml-16"; // shortcut for full abstract
     public static final String LAYOUT_ID_NAACCR_XML_16_ABSTRACT = "naaccr-xml-16-abstract";
     public static final String LAYOUT_ID_NAACCR_XML_16_MODIFIED = "naaccr-xml-16-modified";
     public static final String LAYOUT_ID_NAACCR_XML_16_CONFIDENTIAL = "naaccr-xml-16-confidential";
     public static final String LAYOUT_ID_NAACCR_XML_16_INCIDENCE = "naaccr-xml-16-incidence";
-    public static final String LAYOUT_ID_NAACCR_XML_15 = "naaccr-xml-15"; // abstract for full abstract
+    public static final String LAYOUT_ID_NAACCR_XML_15 = "naaccr-xml-15"; // shortcut for full abstract
     public static final String LAYOUT_ID_NAACCR_XML_15_ABSTRACT = "naaccr-xml-15-abstract";
     public static final String LAYOUT_ID_NAACCR_XML_15_MODIFIED = "naaccr-xml-15-modified";
     public static final String LAYOUT_ID_NAACCR_XML_15_CONFIDENTIAL = "naaccr-xml-15-confidential";
     public static final String LAYOUT_ID_NAACCR_XML_15_INCIDENCE = "naaccr-xml-15-incidence";
-    public static final String LAYOUT_ID_NAACCR_XML_14 = "naaccr-xml-14"; // abstract for full abstract
+    public static final String LAYOUT_ID_NAACCR_XML_14 = "naaccr-xml-14"; // shortcut for full abstract
     public static final String LAYOUT_ID_NAACCR_XML_14_ABSTRACT = "naaccr-xml-14-abstract";
     public static final String LAYOUT_ID_NAACCR_XML_14_MODIFIED = "naaccr-xml-14-modified";
     public static final String LAYOUT_ID_NAACCR_XML_14_CONFIDENTIAL = "naaccr-xml-14-confidential";
     public static final String LAYOUT_ID_NAACCR_XML_14_INCIDENCE = "naaccr-xml-14-incidence";
     //Fixed column layouts
-    public static final String LAYOUT_ID_NAACCR_18 = "naaccr-18"; // abstract for full abstract
+    public static final String LAYOUT_ID_NAACCR_18 = "naaccr-18"; // shortcut for full abstract
     public static final String LAYOUT_ID_NAACCR_18_ABSTRACT = "naaccr-18-abstract";
     public static final String LAYOUT_ID_NAACCR_18_MODIFIED = "naaccr-18-modified";
     public static final String LAYOUT_ID_NAACCR_18_CONFIDENTIAL = "naaccr-18-confidential";
     public static final String LAYOUT_ID_NAACCR_18_INCIDENCE = "naaccr-18-incidence";
-    public static final String LAYOUT_ID_NAACCR_16 = "naaccr-16"; // abstract for full abstract
+    public static final String LAYOUT_ID_NAACCR_16 = "naaccr-16"; // shortcut for full abstract
     public static final String LAYOUT_ID_NAACCR_16_ABSTRACT = "naaccr-16-abstract";
     public static final String LAYOUT_ID_NAACCR_16_MODIFIED = "naaccr-16-modified";
     public static final String LAYOUT_ID_NAACCR_16_CONFIDENTIAL = "naaccr-16-confidential";
     public static final String LAYOUT_ID_NAACCR_16_INCIDENCE = "naaccr-16-incidence";
-    public static final String LAYOUT_ID_NAACCR_15 = "naaccr-15"; // abstract for full abstract
+    public static final String LAYOUT_ID_NAACCR_15 = "naaccr-15"; // shortcut for full abstract
     public static final String LAYOUT_ID_NAACCR_15_ABSTRACT = "naaccr-15-abstract";
     public static final String LAYOUT_ID_NAACCR_15_MODIFIED = "naaccr-15-modified";
     public static final String LAYOUT_ID_NAACCR_15_CONFIDENTIAL = "naaccr-15-confidential";
     public static final String LAYOUT_ID_NAACCR_15_INCIDENCE = "naaccr-15-incidence";
-    public static final String LAYOUT_ID_NAACCR_14 = "naaccr-14"; // abstract for full abstract
+    public static final String LAYOUT_ID_NAACCR_14 = "naaccr-14"; // shortcut for full abstract
     public static final String LAYOUT_ID_NAACCR_14_ABSTRACT = "naaccr-14-abstract";
     public static final String LAYOUT_ID_NAACCR_14_MODIFIED = "naaccr-14-modified";
     public static final String LAYOUT_ID_NAACCR_14_CONFIDENTIAL = "naaccr-14-confidential";
     public static final String LAYOUT_ID_NAACCR_14_INCIDENCE = "naaccr-14-incidence";
-    public static final String LAYOUT_ID_NAACCR_13 = "naaccr-13"; // abstract for full abstract
+    public static final String LAYOUT_ID_NAACCR_13 = "naaccr-13"; // shortcut for full abstract
     public static final String LAYOUT_ID_NAACCR_13_ABSTRACT = "naaccr-13-abstract";
     public static final String LAYOUT_ID_NAACCR_13_MODIFIED = "naaccr-13-modified";
     public static final String LAYOUT_ID_NAACCR_13_CONFIDENTIAL = "naaccr-13-confidential";
     public static final String LAYOUT_ID_NAACCR_13_INCIDENCE = "naaccr-13-incidence";
-    public static final String LAYOUT_ID_NAACCR_12 = "naaccr-12"; // abstract for full abstract
+    public static final String LAYOUT_ID_NAACCR_12 = "naaccr-12"; // shortcut for full abstract
     public static final String LAYOUT_ID_NAACCR_12_ABSTRACT = "naaccr-12-abstract";
     public static final String LAYOUT_ID_NAACCR_12_MODIFIED = "naaccr-12-modified";
     public static final String LAYOUT_ID_NAACCR_12_CONFIDENTIAL = "naaccr-12-confidential";
@@ -115,6 +120,7 @@ public final class LayoutFactory {
 
     // make sure to put all the "aliases" in this list...
     static {
+        _INTERNAL_LAYOUT_ID_ALIASES.put(LAYOUT_ID_NAACCR_XML_22, LAYOUT_ID_NAACCR_XML_22_ABSTRACT);
         _INTERNAL_LAYOUT_ID_ALIASES.put(LAYOUT_ID_NAACCR_XML_21, LAYOUT_ID_NAACCR_XML_21_ABSTRACT);
         _INTERNAL_LAYOUT_ID_ALIASES.put(LAYOUT_ID_NAACCR_XML_18, LAYOUT_ID_NAACCR_XML_18_ABSTRACT);
         _INTERNAL_LAYOUT_ID_ALIASES.put(LAYOUT_ID_NAACCR_XML_16, LAYOUT_ID_NAACCR_XML_16_ABSTRACT);
@@ -134,6 +140,11 @@ public final class LayoutFactory {
     // make sure to add the most recent layouts first, they will be "tried" in that order (important for discovery mechanism)
     static {
         //NAACCR XML
+        _INTERNAL_LAYOUTS.put(LAYOUT_ID_NAACCR_XML_22, "NAACCR XML 21 Abstract");
+        _INTERNAL_LAYOUTS.put(LAYOUT_ID_NAACCR_XML_22_ABSTRACT, "NAACCR XML 22 Abstract");
+        _INTERNAL_LAYOUTS.put(LAYOUT_ID_NAACCR_XML_22_MODIFIED, "NAACCR XML 22 Modified");
+        _INTERNAL_LAYOUTS.put(LAYOUT_ID_NAACCR_XML_22_CONFIDENTIAL, "NAACCR XML 22 Confidential");
+        _INTERNAL_LAYOUTS.put(LAYOUT_ID_NAACCR_XML_22_INCIDENCE, "NAACCR XML 22 Incidence");
         _INTERNAL_LAYOUTS.put(LAYOUT_ID_NAACCR_XML_21, "NAACCR XML 21 Abstract");
         _INTERNAL_LAYOUTS.put(LAYOUT_ID_NAACCR_XML_21_ABSTRACT, "NAACCR XML 21 Abstract");
         _INTERNAL_LAYOUTS.put(LAYOUT_ID_NAACCR_XML_21_MODIFIED, "NAACCR XML 21 Modified");