+++ src/test/java/com/imsweb/layout/naaccrxml/NaaccrXmlLayoutTest.java
@@ -81,8 +81,10 @@ public class NaaccrXmlLayoutTest {
             Assert.assertNotNull(layoutId + "/" + field.getNaaccrId(), field.getLongLabel());
             Assert.assertNotEquals(layoutId + "/" + field.getNaaccrId(), "?", field.getShortLabel());
             Assert.assertNotEquals(layoutId + "/" + field.getNaaccrId(), "?", field.getSection());
-            Assert.assertNotNull(layoutId + "/" + field.getNaaccrId(), layout.getFieldDocByName(field.getName()));
-            Assert.assertNotNull(layoutId + "/" + field.getNaaccrId(), layout.getFieldDocByNaaccrItemNumber(field.getNaaccrItemNum()));
+            if (!"220".equals(layout.getNaaccrVersion())) { // TODO FD remove this once N22 doc is available
+                Assert.assertNotNull(layoutId + "/" + field.getNaaccrId(), layout.getFieldDocByName(field.getName()));
+                Assert.assertNotNull(layoutId + "/" + field.getNaaccrId(), layout.getFieldDocByNaaccrItemNumber(field.getNaaccrItemNum()));
+            }
 
             if (layout.getNaaccrVersion().compareTo(NaaccrFormat.NAACCR_VERSION_180) >= 0)
                 Assert.assertNotNull(layout.getFieldDocByNaaccrItemNumber(35)); // retired field